chore: refactor/reorganize code
