+++ packages/playground/src/components/CustomErrorTab.tsx
