+++ packages/playground/src/components/RainbowImage.module.css
