+++ packages/playground/src/components/RainbowImage.tsx
