+++ packages/playground/src/components/Stars.module.css
