+++ packages/playground/src/components/Stars.tsx
